+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.2.0.0
 
+* Add `dstbRequiredTopLevelGuards` to `TxBody`
+* Add `dstbrRequiredTopLevelGuards` to `DijkstraSubTxBodyRaw`
+* Add `requiredTopLevelGuardsL` to `DijkstraEraTxBody` class
 * Add `DijkstraContextError`
 * Add `dtbSubTransactions` to `TxBody`
 * Add `subTransactionsTxBodyL` method to `DijkstraEraTxBody` class